[ADD] multiple changes

+++ product_variant_custom_sale/models/sale.py
@@ -70,6 +70,29 @@ class SaleOrderLine(models.Model):
             "CHECK(display_type IS NOT NULL OR (product_uom IS NOT NULL))",
             "Missing required fields on accountable sale order line."),
     ]
+
+    def get_product_dict(self, tmpl_id, attributes):
+        values = attributes.mapped("value_id.id")
+        return {
+            'product_tmpl_id': tmpl_id.id,
+            'attribute_value_ids': [(6, 0, values)],
+            'active': tmpl_id.active,
+        }
+
+    def create_product_product(self, template=None, attributes=None):
+        product_dict = self.get_product_dict(self.product_tmpl_id, self.product_attribute_ids)
+        self.product_id = self.env['product.product'].create(product_dict)
+
+    @api.model
+    def create(self, values):
+        product_id = values.get('product_id')
+        if product_id and not values.get('product_tmpl_id'):
+            product = self.env['product.product'].browse(product_id)
+            #attribute_ids = product._get_product_attributes_values_dict()
+            values.update({'product_tmpl_id': product.product_tmpl_id.id})
+            
+        return super().create(values)
+
     def _delete_product_attribute_ids(self):
         delete_values = []
         for value in self.product_attribute_ids:
@@ -92,6 +115,11 @@ class SaleOrderLine(models.Model):
     @api.multi
     @api.onchange('product_tmpl_id')
     def onchange_product_template(self):
+        if self.product_tmpl_id and self.product_id.product_tmpl_id == self.product_tmpl_id:
+            return {'domain': {'product_id':
+                              [('product_tmpl_id', '=',
+                                     self.product_tmpl_id.id)]}}
+
         self.ensure_one()
         self.product_attribute_ids = \
             self._delete_product_attribute_ids()
@@ -112,13 +140,20 @@ class SaleOrderLine(models.Model):
                                      self.product_tmpl_id.id)]}}
         return {'domain': {}}
 
-    @api.onchange('product_id')
-    def onchange_product_id(self):
+    @api.onchange("product_id")
+    def product_id_change(self):
+        p1 = self.product_id
         result = super().product_id_change()
-        self.custom_value_ids = self._delete_custom_lines()
-        self.product_attribute_ids = self._delete_product_attribute_ids()
+        #p1 = self.product_id
+        if not self.product_tmpl_id and self.product_id:
+            self.product_tmpl_id = self.product_id.product_tmpl_id
+            self.product_id = p1
         if self.product_id:
-            product = self.product_id
+            self.custom_value_ids = self._delete_custom_lines()
+            self.product_attribute_ids = self._delete_product_attribute_ids()
+        #if self.product_id:
+            product = p1
+
             self.product_attribute_ids = \
                 product._get_product_attributes_values_dict()
 
@@ -126,6 +161,22 @@ class SaleOrderLine(models.Model):
             version = self.product_id._find_version(self.custom_value_ids)
             self.product_version_id = version
         return result
+
+
+#    @api.onchange('product_id')
+#    def onchange_product_id(self):
+#        result = super().product_id_change()
+#        self.custom_value_ids = self._delete_custom_lines()
+##        self.product_attribute_ids = self._delete_product_attribute_ids()
+#        if self.product_id:
+#            product = self.product_id
+#            self.product_attribute_ids = \
+#                product._get_product_attributes_values_dict()
+
+#            self.custom_value_ids = self._set_custom_lines()
+#            version = self.product_id._find_version(self.custom_value_ids)
+#            self.product_version_id = version
+#        return result
 
     @api.onchange('product_attribute_ids')
     def onchange_product_attributes(self):
